move "node engine" to its own file
